Applying action validity in the correct place so results actually render
