+++ src/CareTogether.Core/Engines/PolicyEvaluation/SharedCalculations.cs
@@ -1,5 +1,4 @@
 ﻿using CareTogether.Resources;
-using CareTogether.Resources.Policies;
 using System;
 using System.Collections.Immutable;
 using System.Linq;
@@ -8,16 +7,6 @@ namespace CareTogether.Engines.PolicyEvaluation
 {
     internal static class SharedCalculations
     {
-        internal static CompletedRequirementInfo ApplyValidityPolicyToCompletedRequirement(
-            EffectiveLocationPolicy policy, CompletedRequirementInfo completed)
-        {
-            var actionDefinition = policy.ActionDefinitions[completed.RequirementName];
-            return completed with
-            {
-                ExpiresAtUtc = actionDefinition.Validity.HasValue ? completed.CompletedAtUtc + actionDefinition.Validity : null
-            };
-        }
-
         public sealed record RequirementCheckResult(bool IsMetOrExempted, DateTime? ExpiresAtUtc);
 
         internal static RequirementCheckResult RequirementMetOrExempted(string requirementName,
